+++ Automate/Framework/Machines/Objects/SeedMakerMachine.cs
@@ -7,13 +7,20 @@ using SObject = StardewValley.Object;
 namespace Pathoschild.Stardew.Automate.Framework.Machines.Objects
 {
     /// <summary>A seed maker that accepts input and provides output.</summary>
+    /// <remarks>See the game's default logic in <see cref="SObject.performObjectDropInAction"/>.</remarks>
     internal class SeedMakerMachine : GenericObjectMachine<SObject>
     {
         /*********
         ** Fields
         *********/
-        /// <summary>A crop ID => seed ID lookup.</summary>
-        private static readonly IDictionary<int, int> CropSeedIDs = SeedMakerMachine.GetCropSeedIDs();
+        /// <summary>A lookup which maps item IDs produced by crops to their seed ID.</summary>
+        private static readonly IDictionary<int, int> SeedLookup = new Dictionary<int, int>();
+
+        /// <summary>The number of defined crops when <see cref="SeedLookup"/> was last updated.</summary>
+        private static int LastCropsCount = -1;
+
+        /// <summary>The game tick when <see cref="SeedLookup"/> was last updated.</summary>
+        private static int LastCacheTick = -1;
 
 
         /*********
@@ -31,13 +38,15 @@ namespace Pathoschild.Stardew.Automate.Framework.Machines.Objects
         /// <returns>Returns whether the machine started processing an item.</returns>
         public override bool SetInput(IStorage input)
         {
+            this.UpdateSeedLookup();
+
             SObject machine = this.Machine;
 
             // crop => seeds
             if (input.TryGetIngredient(this.IsValidCrop, 1, out IConsumable crop))
             {
                 crop.Reduce();
-                int seedID = SeedMakerMachine.CropSeedIDs[crop.Sample.ParentSheetIndex];
+                int seedID = SeedMakerMachine.SeedLookup[crop.Sample.ParentSheetIndex];
 
                 Random random = new Random((int)Game1.stats.DaysPlayed + (int)Game1.uniqueIDForThisGame / 2 + (int)machine.TileLocation.X + (int)machine.TileLocation.Y * 77 + Game1.timeOfDay);
                 machine.heldObject.Value = new SObject(seedID, random.Next(1, 4));
@@ -54,33 +63,42 @@ namespace Pathoschild.Stardew.Automate.Framework.Machines.Objects
 
 
         /*********
-        ** Public methods
+        ** Private methods
         *********/
         /// <summary>Get whether a given item is a crop compatible with the seed marker.</summary>
         /// <param name="item">The item to check.</param>
-        public bool IsValidCrop(ITrackedStack item)
+        private bool IsValidCrop(ITrackedStack item)
         {
             return
                 item.Type == ItemType.Object
                 && item.Sample.ParentSheetIndex != 433 // seed maker doesn't allow coffee beans
-                && SeedMakerMachine.CropSeedIDs.ContainsKey(item.Sample.ParentSheetIndex);
+                && SeedMakerMachine.SeedLookup.ContainsKey(item.Sample.ParentSheetIndex);
         }
 
-        /// <summary>Get a crop ID => seed ID lookup.</summary>
-        public static IDictionary<int, int> GetCropSeedIDs()
+        /// <summary>Update the cached item => seed ID lookup.</summary>
+        /// <remarks>Derived from <see cref="SObject.performObjectDropInAction"/> and <see cref="Crop.InferSeedIndex"/>.</remarks>
+        private void UpdateSeedLookup()
         {
-            IDictionary<int, int> lookup = new Dictionary<int, int>();
-
-            IDictionary<int, string> cropData = Game1.content.Load<Dictionary<int, string>>("Data\\Crops");
-            foreach (KeyValuePair<int, string> entry in cropData)
+            if (Game1.ticks > SeedMakerMachine.LastCacheTick)
             {
-                int dataCropID = Convert.ToInt32(entry.Value.Split('/')[3]);
-                int dataSeedID = entry.Key;
-                if (!lookup.ContainsKey(dataCropID)) // if multiple crops have the same seed, use the earliest one (which is more likely the vanilla seed)
-                    lookup[dataCropID] = dataSeedID;
-            }
+                var cache = SeedMakerMachine.SeedLookup;
+                var crops = Game1.content.Load<Dictionary<int, string>>("Data\\Crops");
+                if (crops.Count != SeedMakerMachine.LastCropsCount)
+                {
+                    cache.Clear();
 
-            return lookup;
+                    foreach (KeyValuePair<int, string> entry in crops)
+                    {
+                        int seedId = entry.Key;
+                        int produceId = Convert.ToInt32(entry.Value.Split('/')[3]);
+                        if (!cache.ContainsKey(produceId)) // use first crop found per game logic
+                            cache[produceId] = seedId;
+                    }
+                }
+
+                SeedMakerMachine.LastCacheTick = Game1.ticks;
+                SeedMakerMachine.LastCropsCount = crops.Count;
+            }
         }
     }
 }